fix: rename env
