+++ yurlungur/core/__init__.py
@@ -2,7 +2,7 @@ from yurlungur.core.app import application
 from yurlungur.core.builder import Builder
 from yurlungur.core.command import file, cmd
 from yurlungur.core.proxy import YObject, YFile, YNode
-from yurlungur.core.enviroment import (
+from yurlungur.core.env import (
     installed, Maya, Houdini, Unreal, Qt,
     MayaBin, HoudiniBin, UnrealBin,
 )
