+++ yurlungur/tool/util.py
@@ -29,8 +29,11 @@ def __db_attr():
     conn = sqlite3.connect(cache)
     c = conn.cursor()
     c.execute('''CREATE TABLE stocks
-                 (date text, trans text, symbol text, qty real, price real)''')
-    c.execute("INSERT INTO stocks VALUES ('2006-01-05','BUY','RHAT',100,35.14)")
+                 (date text, trans text, symbol text, qty real, price real)'''
+              )
+    c.execute(
+        "INSERT INTO stocks VALUES ('2006-01-05','BUY','RHAT',100,35.14)"
+    )
     conn.commit()
     conn.close()
